- Changed the argument name check with arguments array length check places to avoid ArrayIndexOutOfBoundsException - Extract logic if logging into the account was successful and pass   methods reference in ifPresent() method - Remove empty line - Change access modifier of CardGenerator class to default - Move DBConfiguration class to the configuration package
