+++ src/main/java/configuration/DBConfiguration.java
@@ -1,4 +1,4 @@
-package dao;
+package configuration;
 
 import org.sqlite.SQLiteDataSource;
